% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{arrange_times_matrix}
\alias{arrange_times_matrix}
\title{Does something}
\usage{
arrange_times_matrix(ti, tb, ts, tf)
}
\arguments{
\item{ti}{no idea}

\item{tb}{no idea}

\item{ts}{no idea}

\item{tf}{no idea}
}
\value{
result
}
\description{
Does something
}
