% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISSE_functions.R
\name{BISSE_loglik_shift}
\alias{BISSE_loglik_shift}
\title{BISSE loglik}
\usage{
BISSE_loglik_shift(parsM, parsS, brtsM, brtsS)
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}
}
\value{
loglik
}
\description{
Provides BISSE loglik function
}
\author{
Giovanni Laudanno
}
