% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{brts2time_intervals}
\alias{brts2time_intervals}
\title{Brts to time intervals}
\usage{
brts2time_intervals(brts)
}
\arguments{
\item{brts}{branchin times}
}
\value{
The time intervals
}
\description{
Calculates time intervals given the branching times
}
\author{
Giovanni Laudanno
}
