% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{cat2}
\alias{cat2}
\title{cat2}
\usage{
cat2(message, verbose, ...)
}
\arguments{
\item{message}{the message to print}

\item{verbose}{set it to TRUE if you want to see the outputs on screen}
}
\value{
prints on screen
}
\description{
If verbose == TRUE cats the message, otherwise stays silent
}
\author{
Giovanni Laudanno
}
