% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{check_input_data_coherence}
\alias{check_input_data_coherence}
\title{Does something}
\usage{
check_input_data_coherence(dataset, N0)
}
\arguments{
\item{dataset}{dataset to provide to test likelihoods}

\item{N0}{number of lineages at some unknown time}
}
\value{
result
}
\description{
Does something
}
