% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{check_number_of_species}
\alias{check_number_of_species}
\title{Does something}
\usage{
check_number_of_species(lineages)
}
\arguments{
\item{lineages}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
