% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{combine_ddd_nodiv}
\alias{combine_ddd_nodiv}
\title{Combine pn in the old wrong way}
\usage{
combine_ddd_nodiv(lambda, mu, times, tbar, n_max = 100)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{times}{times}

\item{tbar}{time left from shift time to the present}

\item{n_max}{maximum number of lineages to consider}
}
\value{
Convolution of the probabilities for all the processes
}
\description{
Uses the function DDD::conv to convolve all the processes before the shift and
 imposes the death before the present of all species that
 are not visible in the phylogeny
 It doesn't divide by N. Used to check on old models.
}
\author{
Giovanni Laudanno
}
