% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{combine_pns}
\alias{combine_pns}
\title{Combine pn}
\usage{
combine_pns(lambda, mu, times, tbar, n_max = 100, fun = sls::pn_bar)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{times}{times}

\item{tbar}{time left from shift time to the present}

\item{n_max}{maximum number of lineages to consider}

\item{fun}{a function}
}
\value{
Convolution of the probabilities for all the processes
}
\description{
Convolutes all the processes before the shift and imposes the death before the present of all species that are not visible in the phylogeny
}
\author{
Giovanni Laudanno
}
