% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{combine_pns0}
\alias{combine_pns0}
\title{Combine pn (basic)}
\usage{
combine_pns0(lambda, mu, times, tbar, n_max = 100)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{times}{times}

\item{tbar}{time left from shift time to the present}

\item{n_max}{maximum number of lineages to consider}
}
\value{
Convolution of the probabilities for all the processes
}
\description{
Convolution function. Basic but slow.
}
\author{
Giovanni Laudanno
}
