% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_convolutions.R
\name{combine_pns_noratio}
\alias{combine_pns_noratio}
\title{Combine pn in the old wrong way}
\usage{
combine_pns_noratio(lambda, mu, ts, tbar, nmax = 100, fun = sls:::pn_bar)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{tbar}{time left from shift time to the present}

\item{nmax}{maximum number of lineages to consider}
}
\value{
Convolution of the probabilities for all the processes
}
\description{
Convolutes all the processes before the shift and imposes the death before the present of all species that are not visible in the phylogeny. It doesn't divide by N. Used to check on old models.
}
\author{
Giovanni Laudanno
}
