% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{cut_loglik_from_name}
\alias{cut_loglik_from_name}
\title{Cut word "loglik" from a name}
\usage{
cut_loglik_from_name(function_name)
}
\value{
clean name
}
\description{
Cut word "loglik" from a name
}
\author{
Giovanni Laudanno
}
