% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihoods.R
\name{DDD_conditioning}
\alias{DDD_conditioning}
\title{DDD-conditioning}
\usage{
DDD_conditioning(pars, brtsM, brtsS, lx, ddep = 1)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambdaM, i.e. speciation rate of the main clade;
  \item pars[2] is muM, i.e. extinction rate of the main clade;
  \item pars[3] is lambdaS, i.e. speciation rate of the sub clade;
  \item pars[4] is muS, i.e. extinction rate of the sub clade;
}}

\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}

\item{lx}{size of the matrix}

\item{ddep}{specifies the kind of diversity-dependent model you want to use}
}
\value{
Conditioning probability
}
\description{
Calculates the conditioning from the DDD package for the Key Innovation model
}
\author{
Giovanni Laudanno
}
