% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{debug_DDD_conditioning}
\alias{debug_DDD_conditioning}
\title{Does something}
\usage{
debug_DDD_conditioning(lambdas, mus, brtsM, brtsS, lx, ddep = 1)
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}

\item{lx}{size of the matrix}

\item{ddep}{specifies the kind of diversity-dependent model you want to use}
}
\value{
result
}
\description{
Does something
}
