% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_debug.R
\name{debug_Pc_1shift}
\alias{debug_Pc_1shift}
\title{Does something}
\usage{
debug_Pc_1shift(brtsM, brtsS, lambdas, mus)
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}
}
\value{
result
}
\description{
Does something
}
