% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{dft}
\alias{dft}
\title{dft}
\usage{
dft(vec)
}
\arguments{
\item{vec}{a vector or a matrix to be transformed}
}
\value{
dft
}
\description{
Computes the Discrete Fourier Transform (dft)
}
\author{
Giovanni Laudanno
}
