% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_bisse.R
\name{Et}
\alias{Et}
\title{BISSE's E}
\usage{
Et(pars, t0, tf, E0, D0)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambdaM, i.e. speciation rate of the main clade;
  \item pars[2] is muM, i.e. extinction rate of the main clade;
  \item pars[3] is lambdaS, i.e. speciation rate of the sub clade;
  \item pars[4] is muS, i.e. extinction rate of the sub clade;
}}
}
\value{
E(t)
}
\description{
Provides BISSE E's function
}
\author{
Giovanni Laudanno
}
