% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{export_results_to_xls_2}
\alias{export_results_to_xls_2}
\title{Does something}
\usage{
export_results_to_xls_2(Nsims, sim_function, lik_function, result,
  sample_threshold = 1e+05)
}
\arguments{
\item{Nsims}{number of simulation}

\item{sim_function}{simulation function (no idea what that means), can be
\code{\link{sim_custom}} or some unknown others}

\item{lik_function}{likelihood function, can be
\code{\link{lik_custom}} or some unknown others}

\item{result}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
