% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{get_function_names}
\alias{get_function_names}
\title{Get function names}
\usage{
get_function_names(models)
}
\arguments{
\item{models}{the models you want to use to define the likelihood}
}
\value{
function names
}
\description{
Get function names
}
\author{
Giovanni Laudanno
}
