% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{get_model_names}
\alias{get_model_names}
\title{Check if provided models make sense}
\usage{
get_model_names(function_names, verbose = FALSE)
}
\arguments{
\item{function_names}{function names}

\item{verbose}{set it to TRUE if you want to see the outputs on screen}
}
\value{
models names
}
\description{
Check if provided models make sense
}
\author{
Giovanni Laudanno
}
