% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{get_sls_param_names}
\alias{get_sls_param_names}
\title{Get the names of the parameters used in the MBD model}
\usage{
get_sls_param_names()
}
\description{
Get the names of the parameters used in the MBD model
}
\author{
Giovanni Laudanno
}
