% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{get_std2}
\alias{get_std2}
\title{Does something}
\usage{
get_std2(oks, lik_result, sim_result)
}
\arguments{
\item{oks}{internal variable. do not use}

\item{lik_result}{internal variable. do not use}

\item{sim_result}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
