% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{idft}
\alias{idft}
\title{Inverse dft}
\usage{
idft(vec)
}
\arguments{
\item{vec}{a vector or a matrix to be transformed}
}
\value{
Inverse dft
}
\description{
Computes the inverse Discrete Fourier Transform
}
\author{
Giovanni Laudanno
}
