% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{L2LTT}
\alias{L2LTT}
\title{L to Lineage-Through-Time}
\usage{
L2LTT(L, shifts = NULL)
}
\value{
The function returns a matrix:
\itemize{
\item the first row contains the time points when the amount of lineages change;
\item the second row gives the number of lineages immediately after that time point;
}
}
\description{
From an L matrix it yields the corresponding lineage-through-time function.
}
\author{
Giovanni Laudanno
}
