% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihood.R
\name{lik_custom}
\alias{lik_custom}
\title{Custom likelihood equation (for lack of better words)}
\usage{
lik_custom(lambdas, mus, ti, tf, tb, ts, N0 = 1, input_check = 1)
}
\description{
Custom likelihood equation (for lack of better words)
}
