% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihood.R
\name{lik_custom_single_lineage3}
\alias{lik_custom_single_lineage3}
\title{Custom likelihood function for single lineages. You can use it after you apply split_times_matrix.
Does something}
\usage{
lik_custom_single_lineage3(dataset, N0 = 1, input_check = 1)
}
\arguments{
\item{dataset}{dataset to provide to test likelihoods}

\item{N0}{number of lineages at some unknown time}

\item{input_check}{if one (why not TRUE?) the input is
checked (why not always?)}
}
\value{
result
}
\description{
Custom likelihood function for single lineages. You can use it after you apply split_times_matrix.
Does something
}
