% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihood.R
\name{lik_custom_split3}
\alias{lik_custom_split3}
\title{Custom likelihood function that applies the procedure after tree splitting. It uses lik_custom_single_lineage for each lineages.
Does something}
\usage{
lik_custom_split3(lambdas, mus, ti, tf, tb, ts, N0 = 1, input_check = 1)
}
\arguments{
\item{lambdas}{speciation initiation rates}

\item{mus}{extinction rates}

\item{ti}{no idea}

\item{tf}{no idea}

\item{tb}{no idea}

\item{ts}{no idea}

\item{N0}{number of lineages at some unknown time}

\item{input_check}{if one (why not TRUE?) the input is
checked (why not always?)}
}
\value{
result
}
\description{
Custom likelihood function that applies the procedure after tree splitting. It uses lik_custom_single_lineage for each lineages.
Does something
}
