% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihood.R
\name{lik_Pix}
\alias{lik_Pix}
\title{Does something}
\usage{
lik_Pix(lambdas, mus, times)
}
\arguments{
\item{lambdas}{speciation initiation rates}

\item{mus}{extinction rates}
}
\value{
result
}
\description{
Does something
}
