% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihoods.R
\name{lik_shift_DDD0}
\alias{lik_shift_DDD0}
\title{DDD-likelihood (old version)}
\usage{
lik_shift_DDD0(pars, brtsM, brtsS, cond = 0, lx = floor(min(20 *
  max(length(brtsM), length(brtsS)), 1000)), methode = "lsoda")
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambdaM, i.e. speciation rate of the main clade;
  \item pars[2] is muM, i.e. extinction rate of the main clade;
  \item pars[3] is lambdaS, i.e. speciation rate of the sub clade;
  \item pars[4] is muS, i.e. extinction rate of the sub clade;
}}

\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and both crown descendents in the main clade;
}}

\item{lx}{size of the matrix}
}
\value{
The likelihood
}
\description{
Calculates the likelihood calling the routine from the DDD package
}
\author{
Giovanni Laudanno
}
