% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihoods.R
\name{lik_shift_P2}
\alias{lik_shift_P2}
\title{P-likelihood}
\usage{
lik_shift_P2(pars1, pars2 = c(100, 1, 1, brtsM[2], 0, 2), brtsM, brtsS,
  missnumspec = c(0, 0))
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}
}
\value{
The likelihood
}
\description{
Calculates the likelihood convoluting Nee's functions
}
\author{
Giovanni Laudanno
}
