% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihood.R
\name{lik_ut}
\alias{lik_ut}
\title{Does something}
\usage{
lik_ut(lambda, mu, time)
}
\arguments{
\item{lambda}{internal variable. do not use}

\item{mu}{internal variable. do not use}

\item{time}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
