% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_bisse.R
\name{loglik_bisse2}
\alias{loglik_bisse2}
\title{BISSE loglik}
\usage{
loglik_bisse2(pars, brts, N0 = 2, t0 = 0, E0 = 0, D0 = 1,
  LOG = TRUE, lambdaterms = TRUE)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambdaM, i.e. speciation rate of the main clade;
  \item pars[2] is muM, i.e. extinction rate of the main clade;
  \item pars[3] is lambdaS, i.e. speciation rate of the sub clade;
  \item pars[4] is muS, i.e. extinction rate of the sub clade;
}}

\item{brts}{branchin times}
}
\value{
loglik
}
\description{
Provides BISSE loglik function (alternative version)
}
\author{
Giovanni Laudanno
}
