% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_nodiv.R
\name{loglik_bisse_shift}
\alias{loglik_bisse_shift}
\title{BISSE loglik with shift}
\usage{
loglik_bisse_shift(pars_m, pars_s, brts_m, brts_s, cond, n_0 = 2,
  nmax = 100)
}
\arguments{
\item{pars_m}{parameters for the main clade (lambda, mu)}

\item{pars_s}{parameters for the sub clade (lambda, mu)}

\item{brts_m}{branching times for the Main-clade}

\item{brts_s}{branching times for the Sub-clade}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 not available;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
  \item cond = 4 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
}}

\item{n_0}{starting number of lineages}

\item{nmax}{maximum number of lineages to consider}
}
\value{
loglik
}
\description{
Provides BISSE loglik function in a presence of a shift.
}
\author{
Giovanni Laudanno
}
