% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_nodiv.R
\name{loglik_bisse_shift2}
\alias{loglik_bisse_shift2}
\title{BISSE loglik shift}
\usage{
loglik_bisse_shift2(pars, brts, n_0 = 2, t0 = 0, td, LOG = TRUE,
  lambdaterms = TRUE)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambdaM, i.e. speciation rate of the main clade;
  \item pars[2] is muM, i.e. extinction rate of the main clade;
  \item pars[3] is lambdaS, i.e. speciation rate of the sub clade;
  \item pars[4] is muS, i.e. extinction rate of the sub clade;
}}

\item{brts}{branchin times}
}
\value{
loglik
}
\description{
Provides BISSE loglik shift function (alternative version). Yields the old (wrong) BISSE result for the Main Clade only.
}
\author{
Giovanni Laudanno
}
