% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_nodiv.R
\name{loglik_DDD}
\alias{loglik_DDD}
\title{DDD-likelihood}
\usage{
loglik_DDD(pars_m, pars_s, brts_m, brts_s, cond, n_0 = 2, nmax = 100)
}
\arguments{
\item{brts_m}{branching times for the Main-clade}

\item{brts_s}{branching times for the Sub-clade}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and
   on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{nmax}{maximum number of lineages to consider}
}
\value{
The likelihood
}
\description{
Calculates the likelihood calling the routine from the DDD package
}
\author{
Giovanni Laudanno
}
