% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{loglik_preshift2}
\alias{loglik_preshift2}
\title{Calculates the preshift loglik using convolutions}
\usage{
loglik_preshift2(lambdas, mus, brts, n_0 = 2, n_max = 100)
}
\arguments{
\item{lambdas}{speciation rates, for all the clades}

\item{mus}{extinction rate, for all the clades}

\item{brts}{branchin times}

\item{n_0}{starting number of lineages}

\item{n_max}{maximum number of lineages to consider}
}
\value{
Loglik preshift for main clade
}
\description{
Calculates the preshift loglik using convolutions
}
\author{
Giovanni Laudanno
}
