% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{loglik_preshift_unobs}
\alias{loglik_preshift_unobs}
\title{Calculates the preshift loglik using convolutions}
\usage{
loglik_preshift_unobs(lambdas, mus, brts, n_0 = 2, n_max = 100)
}
\arguments{
\item{lambdas}{speciation rates, for all the clades}

\item{mus}{extinction rate, for all the clades}

\item{brts}{branching times provided as a list of vectors, one for the
main clade, one for the subclade}

\item{n_0}{starting number of lineages}

\item{n_max}{maximum number of lineages to consider}
}
\value{
Loglik preshift for main clade and unobserved shift
}
\description{
Calculates the preshift loglik using convolutions as an
 unobserved shift occurs
}
\author{
Giovanni Laudanno
}
