% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_div.R
\name{loglik_sls_p}
\alias{loglik_sls_p}
\title{P-likelihood}
\usage{
loglik_sls_p(pars, brts, cond, n_0 = 2, n_max = 100)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}

\item{brts}{branching times provided as a list of vectors, one for the
main clade, one for the subclade}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{n_0}{starting number of lineages}

\item{n_max}{maximum number of lineages to consider}
}
\value{
The likelihood
}
\description{
Calculates the likelihood convolving Nee's functions
}
\author{
Giovanni Laudanno
}
