% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihoods.R
\name{loglik_slsP}
\alias{loglik_slsP}
\title{P-likelihood}
\usage{
loglik_slsP(pars1, pars2 = c(2 + sum(missnumspec) + 2 * length(brtsM) + 1 + 2
  * length(brtsS), 1, 1, min(abs(brtsM[abs(brtsM) > pars1[7]])), 0, 2), brtsM,
  brtsS, missnumspec = c(0, 0))
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}
}
\value{
The likelihood
}
\description{
Calculates the likelihood convoluting Nee's functions
}
\author{
Giovanni Laudanno
}
