% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_nodiv.R
\name{loglik_slsP_nodiv}
\alias{loglik_slsP_nodiv}
\title{P-likelihood (with no division)}
\usage{
loglik_slsP_nodiv(parsM, parsS, brtsM, brtsS, cond, N0 = 2, nmax = 100)
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and both crown descendents in the main clade;
}}

\item{nmax}{maximum number of lineages to consider}
}
\value{
The likelihood
}
\description{
Calculates the likelihood convoluting Nee's functions. There is no division. It should yield the same likelihood as DDD.
}
\author{
Giovanni Laudanno
}
