% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihoods.R
\name{loglik_slsQ}
\alias{loglik_slsQ}
\title{Q-likelihood}
\usage{
loglik_slsQ(pars1, pars2 = c(100, 1, 1, brtsM[2], 0, 2), brtsM, brtsS)
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}
}
\value{
The likelihood
}
\description{
Calculates the likelihood integrating the Q-equation
}
\author{
Giovanni Laudanno
}
