% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_convolutions.R
\name{omega}
\alias{omega}
\title{Fourier term}
\usage{
omega(N, n, k)
}
\arguments{
\item{n}{number of lineages}

\item{k}{frequencies in discrete fourier transform}
}
\value{
Fourier term
}
\description{
Provides fourier terms: exp((2 * pi * i)/N)
}
\author{
Giovanni Laudanno
}
