% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{p_transition_matrix}
\alias{p_transition_matrix}
\title{Transition matrix builder}
\usage{
p_transition_matrix(lambda, mu, matrix_size)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{matrix_size}{size of the matrix}
}
\value{
The transition matrix
}
\description{
Builds the transition matrix to integrate the differential equations of the P-equation
}
\author{
Giovanni Laudanno
}
