% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{pars_transform_forward}
\alias{pars_transform_forward}
\title{Transform parameters}
\usage{
pars_transform_forward(pars)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}
}
\value{
transformed parameters
}
\description{
Transform parameters according to y = x / (1 + x)
}
\details{
This is not to be called by the user.
}
