% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihoods.R
\name{Pc_1shift0}
\alias{Pc_1shift0}
\title{sls-conditioning (old version)}
\usage{
Pc_1shift0(pars, brtsM, brtsS, Nmax = 100)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambdaM, i.e. speciation rate of the main clade;
  \item pars[2] is muM, i.e. extinction rate of the main clade;
  \item pars[3] is lambdaS, i.e. speciation rate of the sub clade;
  \item pars[4] is muS, i.e. extinction rate of the sub clade;
}}

\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}
}
\value{
Three different conditioning probabilities
}
\description{
Calculates three different kind of conditioning probabilities
}
\author{
Giovanni Laudanno
}
