% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_conditionings.R
\name{Pc_1shift_old}
\alias{Pc_1shift_old}
\title{sls-conditioning}
\usage{
Pc_1shift_old(pars1, pars2, brtsM, brtsS)
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}
}
\value{
Three different conditioning probabilities
}
\description{
Calculates three different kind of conditioning probabilities
}
\author{
Giovanni Laudanno
}
