% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{phase.factor}
\alias{phase.factor}
\title{Fourier term}
\usage{
phase.factor(n_max, n, k)
}
\arguments{
\item{n_max}{maximum number of lineages to consider}

\item{n}{number of lineages}

\item{k}{frequencies in the Discrete Fourier Transform (DFT)}
}
\value{
Fourier term
}
\description{
Provides fourier terms: exp((2 * pi * i)/N)
}
\author{
Giovanni Laudanno
}
