% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{read_results}
\alias{read_results}
\title{Read saved results}
\usage{
read_results(project_folder = NULL)
}
\arguments{
\item{project_folder}{the folder where you want to save data and results}
}
\value{
results
}
\description{
Read saved results
}
\author{
Giovanni Laudanno
}
