% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{shuffle}
\alias{shuffle}
\title{Does something}
\usage{
shuffle(oks, N)
}
\arguments{
\item{oks}{internal variable. do not use}

\item{N}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
