% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{sim_check_id_presence}
\alias{sim_check_id_presence}
\title{Does something}
\usage{
sim_check_id_presence(lineages, id)
}
\arguments{
\item{lineages}{internal variable. do not use}

\item{id}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
