% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_check_ok_condition}
\alias{sim_check_ok_condition}
\title{Does something}
\usage{
sim_check_ok_condition(lineages, Ntips)
}
\arguments{
\item{lineages}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
