% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_custom_repeat}
\alias{sim_custom_repeat}
\title{Main simulation. Repeated Nsims times.}
\usage{
sim_custom_repeat(dataset, Nsims, sim_function = sim_custom, N0 = 1,
  input_check = FALSE)
}
\arguments{
\item{dataset}{dataset to provide to test likelihoods}

\item{Nsims}{number of simulation}

\item{sim_function}{simulation function (no idea what that means), can be
\code{\link{sim_custom}} or some unknown others}

\item{N0}{number of lineages at some unknown time}

\item{input_check}{if one (why not TRUE?) the input is
checked (why not always?)}
}
\description{
Main simulation. Repeated Nsims times.
}
