% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_event_branching}
\alias{sim_event_branching}
\title{Sim Module. Changes lineages matrix according to the branching event.}
\usage{
sim_event_branching(lineages, branching_id)
}
\arguments{
\item{lineages}{internal variable. do not use}

\item{branching_id}{internal variable. do not use}
}
\value{
result
}
\description{
Sim Module. Changes lineages matrix according to the branching event.
}
