% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_get_changing_lineage}
\alias{sim_get_changing_lineage}
\title{Does something}
\usage{
sim_get_changing_lineage(lineages, ids, changing_id)
}
\arguments{
\item{lineages}{internal variable. do not use}

\item{changing_id}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
