% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_get_ids}
\alias{sim_get_ids}
\title{Sim Module. It gets the ids from the lineages matrix.}
\usage{
sim_get_ids(lineages)
}
\arguments{
\item{lineages}{internal variable. do not use}
}
\value{
result
}
\description{
Sim Module. It gets the ids from the lineages matrix.
}
