% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sim_get_motherclade}
\alias{sim_get_motherclade}
\title{Get the motherclade of a given clade}
\usage{
sim_get_motherclade(l_2 = sls::sim_get_standard_l_2(), clade)
}
\arguments{
\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}

\item{clade}{the id of the clade}
}
\value{
the motherclade
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
