% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sim_get_pool}
\alias{sim_get_pool}
\title{Get the pool from the l_0 table}
\usage{
sim_get_pool(l_0)
}
\arguments{
\item{l_0}{the l table, for a single clade}
}
\value{
the pool of species
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
