% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sim_initialize_data_new_clade}
\alias{sim_initialize_data_new_clade}
\title{Initialize data for a new clade}
\usage{
sim_initialize_data_new_clade(
  data,
  clade,
  pars,
  l_2 = sls::sim_get_standard_l_2(),
  l_matrix_size = 10000
)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{clade}{the id of the clade}

\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See \link{sim_get_standard_l_2} for more info.}

\item{l_matrix_size}{the initial length of the l matrix. It will be
increased if necessary}
}
\value{
data for the clade
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
