% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_make_branching}
\alias{sim_make_branching}
\title{Sim Module. Applies the branching to the specified lineage.}
\usage{
sim_make_branching(lineages, branching_lineage, ids)
}
\arguments{
\item{lineages}{internal variable. do not use}
}
\value{
result
}
\description{
Sim Module. Applies the branching to the specified lineage.
}
