% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_make_branching3}
\alias{sim_make_branching3}
\title{Sim Module. Applies the branching to the specified lineage, but even better than 1 and 2 (I hope).}
\usage{
sim_make_branching3(lineages, branching_lineage, ids)
}
\arguments{
\item{lineages}{internal variable. do not use}
}
\value{
result
}
\description{
Sim Module. Applies the branching to the specified lineage, but even better than 1 and 2 (I hope).
}
