% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{sls_check_input}
\alias{sls_check_input}
\title{Check input}
\usage{
sls_check_input(brts_m, brts_s, cond, n_0, n_max)
}
\arguments{
\item{brts_m}{branching times for the Main-clade}

\item{brts_s}{branching times for the Sub-clade}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{n_0}{starting number of lineages}

\item{n_max}{maximum number of lineages to consider}
}
\value{
nothing
}
\description{
It checks the inputs
}
\author{
Giovanni Laudanno
}
