% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_loglik_choosepar.R
\name{sls_loglik_choosepar}
\alias{sls_loglik_choosepar}
\title{Internal MBD function}
\usage{
sls_loglik_choosepar(trparsopt, trparsfix, idparsopt, idparsfix,
  idparsnoshift, pars2, brtsM, brtsS, missnumspec = c(0, 0),
  loglik_function = sls::loglik_slsP)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
