% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_loglik_choosepar.R
\name{sls_loglik_choosepar0}
\alias{sls_loglik_choosepar0}
\title{Internal MBD function}
\usage{
sls_loglik_choosepar0(loglik_function = sls:::lik_shift_P0, trparsopt,
  trparsfix, idparsopt = 1:4, idparsfix = (1:4)[-idparsopt], brtsM, brtsS,
  cond = 1, pars.transform = 0)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
