% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_loglik_choosepar.R
\name{sls_loglik_choosepar2}
\alias{sls_loglik_choosepar2}
\title{Internal MBD function}
\usage{
sls_loglik_choosepar2(trparsopt, trparsfix, idparsopt, idparsfix, idparsnoshift,
  pars2, brtsM, brtsS, pars.transform = 0, missnumspec = c(0, 0),
  fun = sls::lik_shift_P2)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
