% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{sls_loglik_rhs}
\alias{sls_loglik_rhs}
\title{Builds the right hand side of the ODE set}
\usage{
sls_loglik_rhs(t, x, params)
}
\arguments{
\item{t}{time}

\item{x}{vector to integrate}

\item{params}{transition matrix}
}
\description{
Builds the right hand side of the ODE set
}
\details{
This is not to be called by the user.
}
\author{
Giovanni Laudanno
}
