% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{sls_logliks_div}
\alias{sls_logliks_div}
\title{Logliks with division}
\usage{
sls_logliks_div()
}
\value{
loglik functions with division in sls
}
\description{
Get all the loglik functions with division
}
\author{
Giovanni Laudanno
}
