% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_dummy_functions.R
\name{sls_logliks_dummy}
\alias{sls_logliks_dummy}
\title{Logliks with dummy shift}
\usage{
sls_logliks_dummy()
}
\value{
loglik functions with dummy shift
}
\description{
Get all the loglik functions with dummy shift
}
\author{
Giovanni Laudanno
}
