% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{sls_logliks_experiment}
\alias{sls_logliks_experiment}
\title{Logliks for the experiment}
\usage{
sls_logliks_experiment()
}
\value{
loglik functions to use for the experiment
}
\description{
Get the loglik functions to use for the experiment
}
\author{
Giovanni Laudanno
}
