% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{sls_logliks_nodiv}
\alias{sls_logliks_nodiv}
\title{Logliks with no division}
\usage{
sls_logliks_nodiv()
}
\value{
loglik functions with no division in sls
}
\description{
Get all the loglik functions with no division
}
\author{
Giovanni Laudanno
}
