% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_main.R
\name{sls_main}
\alias{sls_main}
\title{Run the full sls routine}
\usage{
sls_main(
  sim_pars,
  cond = 3,
  crown_age = 5,
  shift_time = 2,
  seed,
  start_pars = c(0.2, 0.1, 0.2, 0.1),
  optim_ids = rep(TRUE, length(start_pars)),
  loglik_functions = sls::sls_logliks_div(),
  project_folder = NULL,
  verbose = FALSE,
  max_iterations = 10000
)
}
\arguments{
\item{sim_pars}{parameters of the simulation}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{crown_age}{the age of the phylogeny}

\item{shift_time}{the time of the shift}

\item{seed}{the seed}

\item{start_pars}{parameters to start from for the search of the likelihood
maximum}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{loglik_functions}{the loglik_functions you want to use}

\item{project_folder}{the folder where you want to save data and results}

\item{verbose}{set it to TRUE if you want to see the outputs on screen}

\item{max_iterations}{maximum number of iterations for the optimizer}
}
\description{
Simulate an sls process and infer parameters maximizing the
likelihood(s) for given function(s)
}
\details{
mle inference
}
