% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_ML.R
\name{sls_ML}
\alias{sls_ML}
\title{Internal MBD function}
\usage{
sls_ML(loglik_function = sls:::lik_shift_P, initparsopt, parsfix,
  idparsopt = 1:4, idparsfix = (1:4)[-idparsopt], brtsM, brtsS, cond = 1,
  optimmethod = "simplex", pars.transform = 1, res = 10 * (1 +
  length(c(brtsM, brtsS))), tol = c(0.001, 1e-04, 1e-06), maxiter = 1000 *
  round((1.25)^length(idparsopt)), changeloglikifnoconv = FALSE,
  verbose = TRUE)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
