% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_ML.R
\name{sls_ML0}
\alias{sls_ML0}
\title{Internal MBD function (old)}
\usage{
sls_ML0(loglik_function = sls:::lik_shift_P0, initparsopt, parsfix,
  idparsopt = 1:4, idparsfix = (1:4)[-idparsopt], brtsM, brtsS, cond = 1,
  optimmethod = "simplex", pars.transform = 1, res = 10 * (1 +
  length(c(brtsM, brtsS))), tol = c(0.001, 1e-04, 1e-06), maxiter = 1000 *
  round((1.25)^length(idparsopt)), changeloglikifnoconv = FALSE,
  verbose = TRUE)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
