% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_ml_cluster.R
\name{sls_ml_cluster}
\alias{sls_ml_cluster}
\title{Internal sls function}
\usage{
sls_ml_cluster(seed, sim_pars = c(0.3, 0.1, 0.6, 0.05), t_d = 4.8,
  cond = 1, initparsopt = c(0.5, 0.3, 0.5, 0.3), age = 10,
  optimmethod = "simplex", tolerance = 0.01, fun = sls::loglik_slsP)
}
\arguments{
\item{seed}{the seed}

\item{sim_pars}{parameters of the simulation}

\item{t_d}{decoupling time}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 not available;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
  \item cond = 4 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
}}

\item{initparsopt}{see DDD package}

\item{age}{the age of the phylogeny}

\item{optimmethod}{see DDD package}

\item{tolerance}{see DDD package}

\item{fun}{a function}
}
\description{
Internal sls function.
}
\details{
This is not to be called by the user.
}
