% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_dummy_functions.R
\name{sls_ml_dummy}
\alias{sls_ml_dummy}
\title{sls Maximum Likelihood}
\usage{
sls_ml_dummy(
  loglik_function = sls::loglik_sls_p_dummy,
  brts,
  start_pars = c(0.5, 0.3),
  n_0 = 2,
  cond = 3,
  optim_ids = rep(TRUE, length(start_pars)),
  true_pars = start_pars,
  verbose = TRUE
)
}
\arguments{
\item{loglik_function}{the loglik function you want to use}

\item{brts}{branchin times}

\item{start_pars}{parameters to start from for the search of the likelihood
maximum}

\item{n_0}{starting number of lineages}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{true_pars}{true parameter values when running the ml process.}

\item{verbose}{set it to TRUE if you want to see the outputs on screen}
}
\value{
best parameters
}
\description{
Calculates ML.
}
