% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.adapt_l_matrix_size}
\alias{sls_sim.adapt_l_matrix_size}
\title{Increase the size of the l_0 table if needed}
\usage{
sls_sim.adapt_l_matrix_size(data, clade)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{clade}{the id of the clade}
}
\value{
a bigger l_0 table
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
