% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.cut_l_matrix}
\alias{sls_sim.cut_l_matrix}
\title{Cuts the L table only to the useful rows}
\usage{
sls_sim.cut_l_matrix(L)
}
\arguments{
\item{L}{the l table}
}
\value{
a clean L table
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
