% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.decide_event}
\alias{sls_sim.decide_event}
\title{Determines the event}
\usage{
sls_sim.decide_event(data, clade, deltas, LS)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{clade}{the id of the clade}

\item{deltas}{in the Doob-Gillespie algorithm,
the collection of delta_n and delta_t, which are, respectively,
the change in number of species and
the waiting time for the next event to occur}

\item{LS}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_LS() for more info.}
}
\value{
the event
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
