% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.get_brts}
\alias{sls_sim.get_brts}
\title{Extract the branching times from data}
\usage{
sls_sim.get_brts(data, LS)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{LS}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_LS() for more info.}
}
\value{
the branching times
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
