% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.get_n_0}
\alias{sls_sim.get_n_0}
\title{Initialize n for a new clade}
\usage{
sls_sim.get_n_0(LS = sls::sls_sim.get_standard_LS(), clade)
}
\arguments{
\item{LS}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_LS() for more info.}

\item{clade}{the id of the clade}
}
\value{
n0 for the clade
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
