% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.get_pars}
\alias{sls_sim.get_pars}
\title{Get sls parameters}
\usage{
sls_sim.get_pars(lambdas, mus, Ks = c(Inf, Inf))
}
\arguments{
\item{lambdas}{speciation rates, for all the clades}

\item{mus}{extinction rate, for all the clades}

\item{Ks}{carrying capacities, for all the clades}
}
\value{
sls parameters
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
