% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.get_pool}
\alias{sls_sim.get_pool}
\title{Get the pool from the L table}
\usage{
sls_sim.get_pool(L)
}
\arguments{
\item{L}{the l table}
}
\value{
the pool of species
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
