% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.get_standard_LS}
\alias{sls_sim.get_standard_LS}
\title{Creates the standard LS}
\usage{
sls_sim.get_standard_LS(crown_age = 10, n_0 = 2, shift_time = 4)
}
\arguments{
\item{crown_age}{the age of the phylogeny}

\item{n_0}{starting number of lineages}

\item{shift_time}{the time of the shift}
}
\value{
LS
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
