% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.get_t_0}
\alias{sls_sim.get_t_0}
\title{Initialize t for a new clade}
\usage{
sls_sim.get_t_0(l_2 = sls::sls_sim.get_standard_l_2(), clade)
}
\arguments{
\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}

\item{clade}{the id of the clade}
}
\value{
t_0 for the clade
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
