% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.read_LL}
\alias{sls_sim.read_LL}
\title{Reads LL}
\usage{
sls_sim.read_LL(LL)
}
\arguments{
\item{LL}{the collection of all the l tables, for all the clades}
}
\value{
LL
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
