% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim.R
\name{sls_sim2}
\alias{sls_sim2}
\title{Does something}
\usage{
sls_sim2(pars1, age, soc, cond)
}
\arguments{
\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and both crown descendents in the main clade;
}}
}
\value{
result
}
\description{
Does something
}
