% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim_sample_deltas}
\alias{sls_sim_sample_deltas}
\title{Sample deltas for Doob-Gillespie}
\usage{
sls_sim_sample_deltas(data, clade, pars)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{clade}{the id of the clade}

\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}
}
\value{
delta_n and delta_t
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
