% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{split_times_matrix}
\alias{split_times_matrix}
\title{Does something}
\usage{
split_times_matrix(times_matrix, N0 = 1)
}
\arguments{
\item{N0}{number of lineages at some unknown time}
}
\value{
result
}
\description{
Does something
}
