% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{update_regimes}
\alias{update_regimes}
\title{Does something}
\usage{
update_regimes(t, regime, times_matrix, r, n, N0 = 1)
}
\arguments{
\item{N0}{number of lineages at some unknown time}
}
\value{
result
}
\description{
Does something
}
