% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{check_input_data_coherence}
\alias{check_input_data_coherence}
\title{Does something}
\usage{
check_input_data_coherence(lambdas, mus, ti, tf, tb, ts, N0)
}
\arguments{
\item{lambdas}{speciation initiation rates}

\item{mus}{extinction rates}

\item{ti}{no idea}

\item{tf}{no idea}

\item{tb}{no idea}

\item{ts}{no idea}

\item{N0}{number of lineages at some unknown time}
}
\value{
result
}
\description{
Does something
}
