% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_convolutions.R
\name{combine_pns0}
\alias{combine_pns0}
\title{Combine pn (basic)}
\usage{
combine_pns0(lambda, mu, ts, tbar, nmax = 100)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{tbar}{time left from shift time to the present}

\item{nmax}{maximum number of lineages to consider}
}
\value{
Convolution of the probabilities for all the processes
}
\description{
Convolution function. Basic but slow.
}
\author{
Giovanni Laudanno
}
