% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(lambda, mu, t, n, tbar, pars, brts, brts_m, brts_s,
  nmax, matrix_size, lx, ddep, k, cond, vec)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{t}{time}

\item{n}{number of lineages}

\item{tbar}{time left from shift time to the present}

\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambdaM, i.e. speciation rate of the main clade;
  \item pars[2] is muM, i.e. extinction rate of the main clade;
  \item pars[3] is lambdaS, i.e. speciation rate of the sub clade;
  \item pars[4] is muS, i.e. extinction rate of the sub clade;
}}

\item{brts}{branchin times}

\item{brts_m}{branching times for the Main-clade}

\item{brts_s}{branching times for the Sub-clade}

\item{nmax}{maximum number of lineages to consider}

\item{matrix_size}{size of the matrix}

\item{lx}{size of the matrix}

\item{ddep}{specifies the kind of diversity-dependent model you want to use}

\item{k}{frequencies in discrete fourier transform}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and
   on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{vec}{a vector or a matrix to be transformed}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
