% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(lambdas, mus, ti, tb, ts, tf, N0, Nsims, lik_function,
  sim_function, input_check)
}
\arguments{
\item{lambdas}{speciation initiation rates}

\item{mus}{extinction rates}

\item{ti}{no idea}

\item{tb}{no idea}

\item{ts}{no idea}

\item{tf}{no idea}

\item{N0}{number of lineages at some unknown time}

\item{Nsims}{number of simulation}

\item{lik_function}{likelihood function, can be
\code{\link{lik_custom}} or some unknown others}

\item{sim_function}{simulation function (no idea what that means), can be
\code{\link{sim_custom}} or some unknown others}

\item{input_check}{if one (why not TRUE?) the input is
checked (why not always?)}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
