% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(age, brts, brts_m, brts_s, brts_clade, cond, clade,
  crown_age, d_0, d_0s, data, delta_n, delta_t, deltas, dropextinct, e_0,
  event, final_time, fun, fun1, fun2, function_name, function_names,
  functions_names, k, lambda, lambdas, mu, mus, ks, l_0, l_1, l_2,
  l_matrix_size, l_matrix, log_scale, lambdaterms, loglik_function, lx,
  matrix_size, message, missnumspec, models, n, n_0, n_0s, n_max,
  optim_ids, output, pars, pars_m, pars_s, pars_clade, pars_transformed,
  project_folder, results, shift_time, sim, sim_pars, start_pars, t, t_c,
  t_d, ts, t_0, t_0s, tf, t_f, td, t_p, tds, t_ds, times, tbar, true_pars,
  vec, verbose, ddep, trparsopt, trparsfix, idparsopt, idparsfix,
  idparsnoshift, initparsopt, optimmethod, tolerance, pars2, seed)
}
\arguments{
\item{age}{the age of the phylogeny}

\item{brts}{branchin times}

\item{brts_m}{branching times for the Main-clade}

\item{brts_s}{branching times for the Sub-clade}

\item{brts_clade}{branching times for a single clade}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 not available;
  \item cond = 3 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
  \item cond = 4 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{clade}{the id of the clade}

\item{crown_age}{the age of the phylogeny}

\item{d_0}{starting value for BiSSE's D function}

\item{d_0s}{starting values for BiSSE's D functions}

\item{data}{contains all the information about the simulated process}

\item{delta_n}{in the Doob-Gillespie algorithm,
the change in number of species}

\item{delta_t}{in the Doob-Gillespie algorithm,
the waiting time for the next event to occur}

\item{deltas}{in the Doob-Gillespie algorithm,
the collection of delta_n and delta_t, which are, respectively,
the change in number of species and
the waiting time for the next event to occur}

\item{dropextinct}{TRUE if you want to remove the dead species from the
tree. FALSE otherwise.}

\item{e_0}{starting value for BiSSE's E function}

\item{event}{the event occurring in the simulated process at a given time}

\item{final_time}{the final time that you want to consider for the survival
of the species considered in the l table}

\item{fun}{a function}

\item{fun1}{a function}

\item{fun2}{another function}

\item{function_name}{function name}

\item{function_names}{function names}

\item{functions_names}{function names}

\item{k}{frequencies in the Discrete Fourier Transform (DFT)}

\item{lambda}{speciation rate}

\item{lambdas}{speciation rates, for all the clades}

\item{mu}{extinction rate}

\item{mus}{extinction rate, for all the clades}

\item{ks}{carrying capacities, for all the clades}

\item{l_0}{the l table, for a single clade}

\item{l_1}{the collection of all the l tables, for all the clades}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}

\item{l_matrix_size}{the initial length of the l matrix. It will be
increased if necessary}

\item{l_matrix}{the l table}

\item{log_scale}{set it to TRUE if you desire the output in log form}

\item{lambdaterms}{set it to TRUE if you desire the powers of lambda
in the likelihood}

\item{loglik_function}{the loglik function you want to use}

\item{lx}{size of the matrix}

\item{matrix_size}{size of the matrix}

\item{message}{the message to print}

\item{missnumspec}{number of missing (unseen) species in the phylogeny}

\item{models}{the models you want to use to define the likelihood}

\item{n}{number of lineages}

\item{n_0}{starting number of lineages}

\item{n_0s}{starting number of lineages for all the clades}

\item{n_max}{maximum number of lineages to consider}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{output}{the mle output}

\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}

\item{pars_m}{parameters for the main clade (lambda, mu)}

\item{pars_s}{parameters for the sub clade (lambda, mu)}

\item{pars_clade}{parameters for a single clade}

\item{pars_transformed}{parameters of the likelihood functions, transformed
according to y = x / (1 + x)}

\item{project_folder}{the folder when you want to save data and results}

\item{results}{mle results}

\item{shift_time}{the time of the shift}

\item{sim}{the results of a sim run}

\item{sim_pars}{parameters of the simulation}

\item{start_pars}{parameters to start from for the search of the likelihood
maximum}

\item{t}{time}

\item{t_c}{crown time}

\item{t_d}{decoupling time}

\item{ts}{times}

\item{t_0}{starting time}

\item{t_0s}{starting time for each clade}

\item{tf}{ending time}

\item{t_f}{ending time}

\item{td}{decoupling time}

\item{t_p}{present time}

\item{tds}{decoupling times}

\item{t_ds}{decoupling times}

\item{times}{times}

\item{tbar}{time left from shift time to the present}

\item{true_pars}{true parameter values when running the ml process.}

\item{vec}{a vector or a matrix to be transformed}

\item{verbose}{set it to TRUE if you want to see the outputs on screen}

\item{ddep}{see DDD package}

\item{trparsopt}{see DDD package}

\item{trparsfix}{see DDD package}

\item{idparsopt}{see DDD package}

\item{idparsfix}{see DDD package}

\item{idparsnoshift}{see DDD package}

\item{initparsopt}{see DDD package}

\item{optimmethod}{see DDD package}

\item{tolerance}{see DDD package}

\item{pars2}{see DDD package}

\item{seed}{the seed}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
