% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_convolutions.R
\name{DFT}
\alias{DFT}
\title{DFT}
\usage{
DFT(vec)
}
\arguments{
\item{vec}{a vector or a matrix to be transformed}
}
\value{
DFT
}
\description{
Computes the Discrete Fourier Transform (DFT)
}
\author{
Giovanni Laudanno
}
