% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_bisse.R
\name{Et}
\alias{Et}
\title{BISSE's E}
\usage{
Et(pars, t_0, tf, E0, D0)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda, i.e. speciation rate;
  \item pars[2] is mu, i.e. extinction rate;
}}

\item{t_0}{starting time}

\item{tf}{ending time}

\item{E0}{starting value for BiSSE's E function}

\item{D0}{starting value for BiSSE's D function}
}
\value{
E(t)
}
\description{
Provides BISSE E's function
}
\author{
Giovanni Laudanno
}
