% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_utils.R
\name{get_function_names}
\alias{get_function_names}
\title{Get function names}
\usage{
get_function_names(loglik_functions)
}
\arguments{
\item{loglik_functions}{the loglik_functions you want to use}
}
\value{
function names
}
\description{
Get function names
}
\author{
Giovanni Laudanno
}
