% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_convolutions.R
\name{IDFT}
\alias{IDFT}
\title{Inverse DFT}
\usage{
IDFT(vec)
}
\arguments{
\item{vec}{a vector or a matrix to be transformed}
}
\value{
Inverse DFT
}
\description{
Computes the inverse Discrete Fourier Transform
}
\author{
Giovanni Laudanno
}
