% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihood.R
\name{lik_single_event}
\alias{lik_single_event}
\title{Does something}
\usage{
lik_single_event(lambda, mu, time1, time2)
}
\arguments{
\item{lambda}{internal variable. do not use}

\item{mu}{internal variable. do not use}

\item{time1}{internal variable. do not use}

\item{time2}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
