% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_bisse.R
\name{loglik_bisse2}
\alias{loglik_bisse2}
\title{BISSE loglik}
\usage{
loglik_bisse2(
  pars,
  brts,
  n_0 = 2,
  t_0 = 0,
  e_0 = 0,
  d_0 = 1,
  log_scale = TRUE,
  lambdaterms = TRUE
)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}

\item{brts}{branchin times}

\item{n_0}{starting number of lineages}

\item{t_0}{starting time}

\item{e_0}{starting value for BiSSE's E function}

\item{d_0}{starting value for BiSSE's D function}

\item{log_scale}{set it to TRUE if you desire the output in log form}

\item{lambdaterms}{set it to TRUE if you desire the powers of lambda
in the likelihood}
}
\value{
loglik
}
\description{
Provides BISSE loglik function (alternative version)
}
\author{
Giovanni Laudanno
}
