% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_likelihoods_nodiv.R
\name{loglik_bisse_shift2}
\alias{loglik_bisse_shift2}
\title{BISSE loglik shift}
\usage{
loglik_bisse_shift2(
  pars,
  brts,
  n_0 = 2,
  t_0 = 0,
  t_d,
  log_scale = TRUE,
  lambdaterms = TRUE
)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}

\item{brts}{branching times provided as a list of vectors, one for the
main clade, one for the subclade}

\item{n_0}{starting number of lineages}

\item{t_0}{starting time}

\item{t_d}{decoupling time}

\item{log_scale}{set it to TRUE if you desire the output in log form}

\item{lambdaterms}{set it to TRUE if you desire the powers of lambda
in the likelihood}
}
\value{
loglik
}
\description{
Provides BISSE loglik shift function (alternative version).
 Yields the old (wrong) BISSE result for the Main Clade only.
}
\author{
Giovanni Laudanno
}
