% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_likelihoods.R
\name{loglik_DDD}
\alias{loglik_DDD}
\title{DDD-likelihood}
\usage{
loglik_DDD(pars1, pars2 = c(100, 1, 1, brtsM[2], 0, 2), brtsM, brtsS,
  missnumspec = c(0, 0))
}
\arguments{
\item{brtsM}{branching times for the Main-clade}

\item{brtsS}{branching times for the Sub-clade}
}
\value{
The likelihood
}
\description{
Calculates the likelihood calling the routine from the DDD package
}
\author{
Giovanni Laudanno
}
