% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_B_example}
\alias{sim_B_example}
\title{Simulates Bart example.}
\usage{
sim_B_example(dataset, N0 = 1, input_check = TRUE)
}
\arguments{
\item{dataset}{dataset to provide to test likelihoods}

\item{N0}{number of lineages at some unknown time}

\item{input_check}{if one (why not TRUE?) the input is
checked (why not always?)}
}
\value{
result
}
\description{
Simulates Bart example.
}
