% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_custom}
\alias{sim_custom}
\title{Main simulation, fully customizable.}
\usage{
sim_custom(lambdas, mus, ti, tb, ts, tf, N0 = 1, input_check = TRUE)
}
\arguments{
\item{lambdas}{speciation initiation rates}

\item{mus}{extinction rates}

\item{ti}{no idea}

\item{tb}{no idea}

\item{ts}{no idea}

\item{tf}{no idea}

\item{N0}{number of lineages at some unknown time}

\item{input_check}{if one (why not TRUE?) the input is
checked (why not always?)}
}
\description{
Main simulation, fully customizable.
}
