% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{sim_series}
\alias{sim_series}
\title{Does something}
\usage{
sim_series(lambdas, mus, times)
}
\arguments{
\item{lambdas}{speciation initiation rates}

\item{mus}{extinction rates}

\item{times}{internal variable. do not use}
}
\value{
result
}
\description{
Does something
}
