% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_loglik_choosepar.R
\name{sls_loglik_choosepar}
\alias{sls_loglik_choosepar}
\title{Calculates the likelihood for a subset of parameters}
\usage{
sls_loglik_choosepar(trparsopt, trparsfix, idparsopt, idparsfix,
  idparsnoshift, pars2, brts_m, brts_s, missnumspec = c(0, 0),
  loglik_function = sls::loglik_slsP)
}
\arguments{
\item{brts_m}{branching times for the Main-clade}

\item{brts_s}{branching times for the Sub-clade}

\item{loglik_function}{the loglik function you want to use}
}
\description{
Calculates the likelihood for a subset of parameters
}
\details{
This is not to be called by the user.
}
