% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_main.R
\name{sls_main}
\alias{sls_main}
\title{Run the full sls routine}
\usage{
sls_main(seed, sim_pars, cond, l_2 = sim_get_standard_l_2(crown_age = 5,
  shift_time = 2), start_pars = c(0.2, 0.1, 0.2, 0.1),
  models = sls_logliks_div(), verbose = TRUE)
}
\arguments{
\item{seed}{the seed}

\item{sim_pars}{parameters of the simulation}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 not available;
  \item cond = 3 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
  \item cond = 4 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}

\item{start_pars}{parameters to start from for the search of the likelihood
maximum}

\item{models}{the models you want to use to define the likelihood}

\item{verbose}{set it to TRUE if you want to see the outputs on screen}
}
\description{
Simulate an sls process and infer parameters maximizing the
likelihood(s) for given function(s)
}
\details{
mle inference
}
