% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_ml.R
\name{sls_ML}
\alias{sls_ML}
\title{sls Maximum Likelihood}
\usage{
sls_ML(loglik_function = sls::loglik_slsP, brtsM, brtsS,
  startpars = c(0.5, 0.3, 0.5, 0.3), cond = 1, N0 = 2)
}
\description{
Calculates ML.
}
\details{
best parameters
}
