% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_ml.R
\name{sls_ml}
\alias{sls_ml}
\title{sls Maximum Likelihood}
\usage{
sls_ml(loglik_function = sls::loglik_slsP, brts_m, brts_s,
  startpars = c(0.5, 0.3, 0.5, 0.3), cond = 1, n_0 = 2)
}
\value{
best parameters
}
\description{
Calculates ML.
}
