% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_ml.R
\name{sls_ml}
\alias{sls_ml}
\title{sls Maximum Likelihood}
\usage{
sls_ml(loglik_function = sls::loglik_sls_p, brts_m, brts_s,
  startpars = c(0.5, 0.3, 0.5, 0.3), cond = 3, n_0 = 2,
  verbose = 1)
}
\arguments{
\item{loglik_function}{the loglik function you want to use}

\item{brts_m}{branching times for the Main-clade}

\item{brts_s}{branching times for the Sub-clade}

\item{startpars}{parameters to start from for the search of the likelihood
maximum}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 not available;
  \item cond = 3 conditions on the survival of the subclade and
  both crown descendents in the main clade;
  \item cond = 4 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
}}

\item{n_0}{starting number of lineages}

\item{verbose}{set it to TRUE if you want to see the outputs on screen}
}
\value{
best parameters
}
\description{
Calculates ML.
}
