% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim_utils.R
\name{sls_sim.check_survival}
\alias{sls_sim.check_survival}
\title{Check if a clade survives untile final_time}
\usage{
sls_sim.check_survival(l_0, final_time = 0)
}
\arguments{
\item{l_0}{the l table, for a single clade}

\item{final_time}{the final time that you want to consider for the survival
of the species considered in the l table}
}
\value{
a boolean
}
\description{
sls_sim module
}
\author{
Giovanni Laudanno
}
