% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_sim.R
\name{sls_sim}
\alias{sls_sim}
\title{Does something}
\usage{
sls_sim(pars, t0, starting_species = c(2, 1), cond)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambdaM, i.e. speciation rate of the main clade;
  \item pars[2] is muM, i.e. extinction rate of the main clade;
  \item pars[3] is lambdaS, i.e. speciation rate of the sub clade;
  \item pars[4] is muS, i.e. extinction rate of the sub clade;
}}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 conditions on the survival of subclade and on the other crown descendents in the main clade;
  \item cond = 3 conditions on the survival of the subclade and both crown descendents in the main clade;
}}
}
\value{
result
}
\description{
Does something
}
