% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_likelihood_formula.R
\name{test_likelihood_formula}
\alias{test_likelihood_formula}
\title{Check the agreement between a likelihood function and a simulation function on a given dataset.}
\usage{
test_likelihood_formula(dataset, N0 = 1, Nsims = 1e+05,
  lik_function = lik_custom, sim_function = sim_custom,
  input_check = TRUE)
}
\arguments{
\item{dataset}{dataset to provide to test likelihoods}

\item{N0}{number of lineages at some unknown time}

\item{Nsims}{number of simulation}

\item{lik_function}{likelihood function, can be
\code{\link{lik_custom}} or some unknown others}

\item{sim_function}{simulation function (no idea what that means), can be
\code{\link{sim_custom}} or some unknown others}

\item{input_check}{if one (why not TRUE?) the input is
checked (why not always?)}
}
\value{
no idea
}
\description{
Check the agreement between a likelihood function and a simulation function on a given dataset.
}
\author{
Giovanni Laudanno
}
