% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_likelihood_formula.R
\name{test_likelihood_formula}
\alias{test_likelihood_formula}
\title{RJCB: I think this should move to the tests folder}
\usage{
test_likelihood_formula(lambdas, mus, ti, tb, ts, tf, N0 = 1, Nsims = 1e+05,
  lik_function = lik_custom, sim_function = sim_custom, input_check = 1)
}
\description{
RJCB: I think this should move to the tests folder
}
