% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_likelihood_formula.R
\name{test_likelihood_formula2}
\alias{test_likelihood_formula2}
\title{Does something}
\usage{
test_likelihood_formula2(Nsims, dataset, lik_function = lik_custom,
  sim_function = sim_custom)
}
\arguments{
\item{Nsims}{number of simulation}

\item{lik_function}{likelihood function, can be
\code{\link{lik_custom}} or some unknown others}

\item{sim_function}{simulation function (no idea what that means), can be
\code{\link{sim_custom}} or some unknown others}
}
\value{
result
}
\description{
Does something
}
