% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_likelihood_formula.R
\name{test_likelihood_formula2}
\alias{test_likelihood_formula2}
\title{Does something}
\usage{
test_likelihood_formula2(s, Nsims)
}
\arguments{
\item{Nsims}{number of simulation}
}
\value{
result
}
\description{
Does something
}
